requirements/maintainer.txt
coverage==3.7.1
flake8==2.3.0
nose-cov==1.6
nose==1.3.4
rednose==0.4.1
sphinx==1.2.3
tox==1.8.1
djangorestframework==3.0.5
sphinx_rtd_theme==0.1.6
